TODO: Support backquote syntactic sugar
Scheme specifies that this should expand into quasiquote/unquote.
http://www.schemers.org/Documents/Standards/R5RS/HTML/r5rs-Z-H-7.html#%_sec_4.2.6

Quote operator in lisp
In lisp, (a 0 1) can both be code (symbol a and integers 0 and 1) and can also
be a list. The quote operator gives you a literal list.
Lisp in general has a notion of special forms, where the arguments are not always
evaluated. This seems like that.

Quote is a compile time operator in that the quoted form is defined and is frozen
at compile time.

Quasiquote operator:
Just like quote. However, this has an unquote operator which allows you to unquote
part of the code. This allows you to parametrize the quoted data structure.
The @ operator allows an unquoted version of a list to come through. If not used,
then unquoted list would remain a list.

Quasiquote is a runtime operator in that the quasiquote form is defined at
compile time, and the unquoted dereferenced at runtime. Almost exactly like
printf, but quasiquoted forms are available everywhere.

Implementation
Our quote implementation converts '(a b c) to List [Atom "quote", List [A, B, C]].
Similarly, we need another atom, unquote which handles the unquote operator.
The unquote-splice-list is also another operator. Runtime again because we do
not have the list to be spliced in until then.